0.0.6 PR

+++ download_toolbox/dataset.py
@@ -43,6 +43,7 @@ class DatasetConfig(DataCollection):
 
     def __init__(self,
                  *,
+                 config_type: str = "dataset_config",
                  existing_dates: list = None,
                  frequency: object = Frequency.DAY,
                  levels: object = (),
@@ -55,7 +56,7 @@ class DatasetConfig(DataCollection):
                  var_files: dict = None,
                  var_names: object = (),
                  **kwargs) -> None:
-        super(DatasetConfig, self).__init__(config_type="dataset_config",
+        super(DatasetConfig, self).__init__(config_type=config_type,
                                             path_components=[frequency.name.lower(), location.name]
                                             if path_components is None else path_components,
                                             **kwargs)
@@ -100,8 +101,6 @@ class DatasetConfig(DataCollection):
             append = []
 
         data_var_path = os.path.join(self.path if not root else self.root_path, *[var, *append])
-        # logging.debug("Handling {}data var path: {}".
-        #               format("root " if root else "", data_var_path))
 
         if not os.path.exists(data_var_path):
             if not missing_error:
@@ -130,13 +129,17 @@ class DatasetConfig(DataCollection):
         for var_name in self.var_files.keys():
             old_files = self.var_files[var_name]
             new_files = [var_file.replace(old_path, self.path) for var_file in old_files]
+            invalid_files = []
 
             for src, dest in zip(old_files, new_files):
-                os.makedirs(os.path.dirname(dest), exist_ok=True)
-                logging.debug("Copying {} to {}".format(src, dest))
-                shutil.copy(src, dest)
-
-            self.var_files[var_name] = new_files
+                if os.path.exists(src):
+                    logging.debug("Copying {} to {}".format(src, dest))
+                    os.makedirs(os.path.dirname(dest), exist_ok=True)
+                    shutil.copy(src, dest)
+                else:
+                    logging.warning("Encountered reference to non-existent data: {}".format(src))
+                    invalid_files.append(dest)
+            self.var_files[var_name] = [fn for fn in new_files if fn not in invalid_files]
 
     def filter_extant_data(self,
                            var_config: VarConfig,
@@ -208,11 +211,11 @@ class DatasetConfig(DataCollection):
             var_names = [v.name for v in self.variables]
 
         logging.debug("Finding files for {}".format(", ".join(var_names)))
-        var_files = [var_filepaths
+        var_files = [var_filepath
                      for vn in var_names
-                     for var_filepaths in self.var_files[vn]]
+                     for var_filepath in self.var_files[vn]
+                     if os.path.exists(var_filepath)]
         logging.info("Got {} filenames to open dataset with!".format(len(var_files)))
-        logging.debug(pformat(var_files))
 
         # TODO: where's my parallel mfdataset please!?
         with dask.config.set(**{'array.slicing.split_large_chunks': True,
@@ -273,8 +276,12 @@ class DatasetConfig(DataCollection):
         if var_filter_list is not None:
             ds = ds.drop_vars(var_filter_list, errors="ignore")
 
-        # TODO: Reduce spatially to required location
-        #  this will also need to set our shape details
+        # TODO: Reduce spatially to required location this will also need to set our shape details
+        # TODO: ideally we should have a broader cache that would allow us to reuse data
+        # TODO: we CANNOT handle nav_lon and nav_lat (e.g. 2D point meshes) yet
+        #if all([f in ds.coords for f in ["latitude", "longitude"]]):
+        #    ds = ds.sel(latitude=slice(self.location.bounds[0], self.location.bounds[2]),
+        #                longitude=slice(self.location.bounds[1], self.location.bounds[3]))
 
         # Reduce temporally to required resolution
         # TODO: Note, https://github.com/pydata/xarray/issues/364 for Grouper functionality?
@@ -345,6 +352,7 @@ class DatasetConfig(DataCollection):
     def var_filepaths(self,
                       var_config: VarConfig,
                       date_batch: list,
+                      file_extension: str = "nc",
                       single_only: bool = False) -> list:
         """
 
@@ -354,7 +362,10 @@ class DatasetConfig(DataCollection):
         :return:
         """
         output_filepaths = list(set([
-            os.path.join(var_config.path, "{}.nc".format(date.strftime(self._output_group_by.date_format)))
+            os.path.join(var_config.path, "{}.{}".format(
+                date.strftime(self._output_group_by.date_format),
+                file_extension
+            ))
             for date in date_batch]))
 
         if len(output_filepaths) > 1 and single_only:
@@ -371,18 +382,6 @@ class DatasetConfig(DataCollection):
 
         return output_filepaths
 
-    #@property
-    #def config(self):
-    #    if self._config is None:
-    #        config_ident = ".".join(self.path_components)
-    #
-    #        logging.debug("Creating dataset configuration with {}".format(config_ident))
-    #        self._config = Configuration(
-    #            config_path=self.config.output_path if self.config.output_path is None else self.root_path,
-    #            config_type=self.config_type,
-    #            identifier=config_ident)
-    #    return self._config
-
     @property
     def frequency(self):
         return self._frequency
@@ -396,12 +395,15 @@ class DatasetConfig(DataCollection):
         for var_name, levels in zip(self._var_names, self._levels):
             for level in levels if levels is not None else [None]:
                 var_config = self.var_config(var_name, level)
-                # logging.debug("Returning configuration: {}".format(var_config))
                 yield var_config
 
     @property
     def var_files(self):
         return self._var_files
+
+    @property
+    def var_prefixes(self):
+        return self._var_names
 
     @var_files.setter
     def var_files(self, value: dict):
